fetch expiring info
