+++ app/controllers/contents_controller.rb
@@ -1,6 +1,13 @@
+
 class ContentsController < ApplicationController
 
   def index
+    @service =FetchTitles.new
+
+    @countries = @service.getCountries
+    country = current_user.country
+
+    @content = @service.getExpiringContent(country)
 
   end
 
@@ -9,6 +16,5 @@ class ContentsController < ApplicationController
 
   def edit
   end
-
 
 end
