+++ app/views/contents/index.html.erb
@@ -1,2 +1,15 @@
 <h1>Contents#index</h1>
 <p>Find me in app/views/contents/index.html.erb</p>
+
+
+<div>
+ <% @content.each do |cont| %>
+ <p>
+ <%= cont[:netflixid] %> - <%= cont[:title] %>
+ <%= cont[:plot] %> - <%= cont[:category] %> - <%= cont[:expiration_date] %>
+ <img src=" <%= cont[:image_url] %>" />
+</p>
+ <% end %>
+</div>
+
+
