+++ app/views/pages/home.html.erb
@@ -1,13 +1,2 @@
 <h1>Pages#home</h1>
 <p>Find me in app/views/pages/home.html.erb</p>
-
-<div>
- <% @countries.each do |country| %>
- <p>
- <%= country[:code] %> - <%= country[:name] %>
-</p>
- <% end %>
-
-
-
-</div>
